Nicer GlobalTargetInfo validation
Changed lookTarget validation on MasterArchives to execute at least once per day (and/or on saving/autosaving).

The mod now checks (once per frame) if there is a MasterArchive queued for validation. All archives are queued once when the user saves/autosaves, and once per day if the user hasn't saved until then.